test(wtr): move base config into shared dir
